login functionality working to homepage
