+++ response_operations_ui/templates/home.html
@@ -1,0 +1,21 @@
+{% extends "layouts/base.html" %}
+
+{% block page_title %} Home | Survey Data Collection {% endblock %}
+
+
+{% block main %}
+
+<div class="container page__container">
+  <div class="grid">
+    <div class="grid__col col-12@m">
+      <div role="main" id="main" class="page__main">
+        <ul class="u-pt-l">
+          <li><a name="surveys-link" href="/surveys/">View list of business surveys</a></li>
+          <li><a name="logout-link" href="/logout">Logout</a></li>
+        </ul>
+      </div>
+    </div>
+  </div>
+</div>
+
+{% endblock main %}
