+++ response_operations_ui/templates/sign_in.html
@@ -4,7 +4,7 @@
 
 
 {% block main %}
-<br>
+
 <div class="container page__container">
   <div class="grid">
     <div class="grid__col col-12@m">
@@ -19,27 +19,10 @@
           {{ form.password(size=20) }} <br>
         </div>
         <br>
-        <input type="submit" class="btn btn--primary" value="Login">
+        <input type="submit" class="btn btn--primary" value="Sign in">
       </form>
     </div>
   </div>
 </div>
-
-
-
-<div class="container page__container">
-  <div class="grid">
-    <div class="grid__col col-12@m">
-      <div role="main" id="main" class="page__main">
-        <ul class="u-pt-l">
-          <li><a name="surveys-link" href="/surveys/">View list of business surveys</a></li>
-          <li><a name="logout-link" href="/logout">Logout</a></li>
-        </ul>
-      </div>
-    </div>
-  </div>
-</div>
-
-
 
 {% endblock main %}
