+++ response_operations_ui/views/collection_exercise.py
@@ -1,16 +1,18 @@
 import logging
 
-from flask import render_template, request
+from flask import Blueprint, render_template, request
 from structlog import wrap_logger
 
-from response_operations_ui import app
 from response_operations_ui.controllers import collection_exercise_controllers
 from response_operations_ui.controllers import collection_instrument_controllers
 
 logger = wrap_logger(logging.getLogger(__name__))
 
+collection_exercise_bp = Blueprint('collection_exercise_bp', __name__,
+                                   static_folder='static', template_folder='templates')
 
-@app.route('/surveys/<short_name>/<period>', methods=['GET'])
+
+@collection_exercise_bp.route('/surveys/<short_name>/<period>', methods=['GET'])
 def view_collection_exercise(short_name, period):
     ce_details = collection_exercise_controllers.get_collection_exercise(short_name, period)
     return render_template('collection-exercise.html', survey=ce_details['survey'],
@@ -18,7 +20,7 @@ def view_collection_exercise(short_name, period):
                            collection_instruments=ce_details['collection_instruments'])
 
 
-@app.route('/surveys/<short_name>/<period>', methods=['POST'])
+@collection_exercise_bp.route('/surveys/<short_name>/<period>', methods=['POST'])
 def upload_collection_instrument(short_name, period):
     error = _validate_collection_instrument()
     ci_loaded = False
